requirements.txt
alpaca-trade-api 
yfinance
backtrader
fastapi 
uvicorn
streamlit
scikit-learn
twilio
python-dotenv